Migrated distribution analysis to its own class. Fixed wrong periods used in analysis. Intervals in analysis are now calculated according to Sturges' rule.
