+++ main.py
@@ -1,6 +1,7 @@
 from nbp_api_caller import NBPApiCaller
 from datetime import datetime, timedelta
 from rate_analysis import RateAnalysis
+from distribution_analysis import DistributionAnalysis
 import common
 
 caller = NBPApiCaller()
@@ -10,8 +11,6 @@ data_range = "1 year"
 
 end_date = datetime.now()
 start_date = end_date - timedelta(days=common.DATA_RANGES[data_range])
-report_type = "monthly"  # Accepted values: "monthly", "quarterly"
-intervals = 4
 
 raw_data = caller.fetch_exchange_rate("USD", "PLN", start_date, end_date)
 
@@ -20,7 +19,6 @@ print(raw_data)
 analysis = RateAnalysis(raw_data)
 statistics = analysis.calculate_statistics()
 sessions = analysis.calculate_sessions()
-distribution = analysis.calculate_changes_distribution(report_type=report_type, intervals=intervals)
 
 print("Statistical Indicators:")
 for key, value in statistics.items():
@@ -30,6 +28,25 @@ print("\nSession Counts:")
 for key, value in sessions.items():
     print(f"  {key}: {value}")
 
-print("\nDistribution of currency value changes:")
-for key, value in distribution.items():
-    print(f"  {key}: {value}")
+start_date_distribution = datetime(2025, 2, 6)
+data_range_distribution = "1 quarter"
+
+if data_range_distribution == "1 quarter":
+    report_type = common.DEFAULT_CURRENCY_CHANGES_HISTOGRAM_REPORT_TYPE 
+else:
+    report_type = common.CURRENCY_CHANGES_HISTOGRAM_REPORT_TYPES[1]
+
+
+end_date_distribution = start_date_distribution + timedelta(days=common.DATA_RANGES[data_range_distribution])
+
+if end_date_distribution > datetime.now():
+    raise ValueError("End date cannot be in the future.")
+else:
+    raw_data_distribution = caller.fetch_exchange_rate("USD", "PLN", start_date_distribution, end_date_distribution)
+
+    distribution_analysis = DistributionAnalysis(raw_data_distribution)
+    distribution = distribution_analysis.calculate_changes_distribution(report_type)
+
+    print("\nDistribution of currency value changes:")
+    for key, value in distribution.items():
+        print(f"  {key}: {value}")
