+++ rate_analysis.py
@@ -1,6 +1,4 @@
 from statistics import median, mode, stdev, mean
-from datetime import datetime
-import numpy as np
 
 class RateAnalysis:
     def __init__(self, data):
@@ -32,52 +30,3 @@ class RateAnalysis:
             'session_downward': downward
         }
         return sessions
-
-    def calculate_periodic_changes(self, report_type):
-        """
-        Calculates currency value changes over monthly or quarterly periods.
-
-        Args:
-            report_type (str): "monthly" or "quarterly".
-
-        Returns:
-            list: A list of value changes for the selected periods.
-        """
-        if report_type not in ["monthly", "quarterly"]:
-            raise ValueError("Invalid report type. Use 'monthly' or 'quarterly'.")
-
-        changes = []
-        rates_by_date = {entry["date"]: entry["rate"] for entry in self.data}
-        dates = sorted(rates_by_date.keys(), key=lambda d: datetime.strptime(d, "%Y-%m-%d"))
-
-        step = 30 if report_type == "monthly" else 91
-
-        for i in range(0, len(dates) - step, step):
-            start_date = dates[i]
-            end_date = dates[i + step]
-            change = rates_by_date[end_date] - rates_by_date[start_date]
-            changes.append(change)
-
-        return changes
-
-    def calculate_changes_distribution(self, report_type, intervals):
-        """
-        Calculates the histogram distribution of currency value changes for monthly/quarterly periods.
-
-        Args:
-            report_type (str): "monthly" or "quarterly".
-            intervals (int): Number of histogram intervals.
-
-        Returns:
-            dict: A dictionary with bins and their respective frequencies.
-        """
-        changes = self.calculate_periodic_changes(report_type=report_type)
-        min_change, max_change = min(changes), max(changes)
-        bins = np.linspace(min_change, max_change, intervals + 1)
-
-        histogram, _ = np.histogram(changes, bins=bins)
-
-        return {
-            "bins": bins.tolist(),
-            "frequency": histogram.tolist()
-        }
